refactor extract class + temperature probe interval for view + url mappings

+++ web/cpd/grails-app/domain/com/mallotore/configuration/Temperature.groovy
@@ -2,7 +2,7 @@ package com.mallotore.configuration
 
 class Temperature {
 	int id
-    long probeIntervalInSeconds
+    int probeIntervalInSeconds
 
     static mapping =  {
 		version false
